requirements.txt
APScheduler==3.6.3
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
chartkick==0.5.0
Django==3.0.3
django-crispy-forms==1.9.0
django-filter==2.2.0
django-multiselectfield==0.1.12
django-pivot==1.8.0
django-select-multiple-field==0.4.2
django-tables2==2.2.1
idna==2.8
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
packaging==20.1
pandas==1.0.1
Pillow==7.0.0
psycopg2==2.8.4
pyparsing==2.4.6
python-dateutil==2.8.1
python-form==0.2.3
pytz==2019.3
PyYAML==5.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
tornado==6.0.3
tzlocal==2.0.0
urllib3==1.25.8
